refactor: reduce main complexity
Refactored main.ts to reduce complexity.
