+++ backend/vizzy-backend/package.json
@@ -31,7 +31,6 @@
     "@nestjs/throttler": "^6.4.0",
     "@supabase/supabase-js": "^2.49.1",
     "@types/nodemailer": "^6.4.17",
-    "cookie-parser": "^1.4.7",
     "express": "^5.0.1",
     "ioredis": "^5.6.0",
     "multer": "^1.4.5-lts.2",
