+++ backend/vizzy-backend/src/main.ts
@@ -1,62 +1,22 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { GlobalHttpExceptionFilter } from './common/filters/http-exception.filter';
-import { ValidationExceptionFilter } from './common/filters/validation-exception.filter';
-import { ZodValidationPipe } from 'nestjs-zod';
-import * as cookieParser from 'cookie-parser';
-import { VersioningType } from '@nestjs/common';
-import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { CustomThrottlerGuard } from './common/guards/throttler.guard';
+import { setupCors } from './setup/setup-cors';
+import { setupEnvironmentSpecificGuards } from './setup/setup-environment-guards';
+import { setupGlobalFilters } from './setup/setup-global-filters';
+import { setupGlobalPipes } from './setup/setup-global-pipes';
+import { setupLogger } from './setup/setup-logger';
+import { setupSwagger } from './setup/setup-swagger-docs';
+import { setupVersioning } from './setup/setup-versioning';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  app.useLogger(app.get(WINSTON_MODULE_NEST_PROVIDER));
-  app.enableVersioning({
-    type: VersioningType.URI,
-    defaultVersion: '1',
-  });
-
-  app.useGlobalFilters(
-    new GlobalHttpExceptionFilter(),
-    new ValidationExceptionFilter(),
-  );
-
-  app.useGlobalPipes(new ZodValidationPipe());
-
-  app.use(cookieParser());
-
-  const config = new DocumentBuilder()
-    .setTitle('Vizzy API')
-    .setDescription('The Vizzy API documentation')
-    .setVersion('1.0')
-    .addBearerAuth()
-    .build();
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('api', app, document);
-
-  app.enableCors({
-    origin: true,
-    methods: 'GET,HEAD,PUT,PATCH,POST,DELETE,OPTIONS',
-    credentials: true,
-    allowedHeaders: 'Content-Type,Accept,Authorization,X-Requested-With',
-    preflightContinue: false,
-    optionsSuccessStatus: 204,
-  });
-
-  const isProduction = process.env.NODE_ENV === 'production';
-
-  if (isProduction) {
-    // Only apply the guard globally in production environments
-    app.useGlobalGuards(app.get(CustomThrottlerGuard));
-    console.log(
-      'Production environment detected: Applying global rate limiter.',
-    );
-  } else {
-    console.log(
-      `Non-production environment detected (${process.env.NODE_ENV}): Skipping global rate limiter.`,
-    );
-  }
+  setupLogger(app);
+  setupVersioning(app);
+  setupGlobalFilters(app);
+  setupGlobalPipes(app);
+  setupSwagger(app);
+  setupCors(app);
+  setupEnvironmentSpecificGuards(app);
 
   await app.listen(process.env.PORT ?? 5000);
 }
