+++ backend/vizzy-backend/src/setup/setup-global-filters.ts
@@ -1,0 +1,9 @@
+import { GlobalHttpExceptionFilter } from '@/common/filters/http-exception.filter';
+import { ValidationExceptionFilter } from '@/common/filters/validation-exception.filter';
+
+export function setupGlobalFilters(app) {
+  app.useGlobalFilters(
+    new GlobalHttpExceptionFilter(),
+    new ValidationExceptionFilter(),
+  );
+}
