+++ backend/vizzy-backend/src/setup/setup-global-pipes.ts
@@ -1,0 +1,5 @@
+import { ZodValidationPipe } from 'nestjs-zod';
+
+export function setupGlobalPipes(app) {
+  app.useGlobalPipes(new ZodValidationPipe());
+}
