+++ backend/vizzy-backend/src/setup/setup-logger.ts
@@ -1,0 +1,6 @@
+import { INestApplication } from '@nestjs/common';
+import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
+
+export function setupLogger(app: INestApplication) {
+  app.useLogger(app.get(WINSTON_MODULE_NEST_PROVIDER));
+}
