+++ backend/vizzy-backend/src/setup/setup-versioning.ts
@@ -1,0 +1,8 @@
+import { VersioningType } from '@nestjs/common/enums';
+
+export function setupVersioning(app) {
+  app.enableVersioning({
+    type: VersioningType.URI,
+    defaultVersion: '1',
+  });
+}
